fix(src): resolve int overflow and log fib values
Also prints the unit behind the duration a calculation took

+++ src/functions.cpp
@@ -1,6 +1,6 @@
 #include "functions.h"
 
-int fibonacci (int index) {
+unsigned long long fibonacci (unsigned long long index) {
     return (index <= 1) ? index : fibonacci (index - 1) + fibonacci (index - 2);
 }
